Adds application name to mail templates
Introduces the application name to the email templates.

This change ensures that the application name is dynamically included in the account creation and password recovery emails, providing users with more context and clarity.

The application name is now configurable in the FancyAdminExtension and passed to the Administration class. It is then accessible in the MailerTrait and rendered in the email templates.

Removes old email templates and refactors existing ones.