Implements lost password functionality.
Adds the functionality for users to reset their password
if they have forgotten it. It introduces new forms
and factories for lost password and new password processes.
It also modifies the sign-in form to include a link to the
lost password page. The changes include injecting the
fancy admin service and the onetime token query factory.