Adds context to authentication process
Adds a context parameter to the authentication process, allowing for more granular control and differentiation of user logins.

This change introduces a 'context' configuration option to the FancyAdminExtension, which is then passed to the FancyAdmin model. The context is utilized during authentication within the SignInFormTrait and SignPresenterTrait, enabling the system to distinguish between different authentication scenarios.