Refactors new password form creation
Changes the creation of the new password form to use a factory without direct dependency injection.
This decouples the form from the identity object, allowing for a more flexible approach where the identity is set after form creation.
Also, it correctly retrieves the object class name for the onetime token to prevent potential issues.